Allow customizing reporting paths
The output formats are currently a mutable global var.

I'd prefer to add a pointer to an ExportFormats to the GraphiteConfig struct
(with fallback to a default if nil), but adding the field would potentially break
backwards compat for users constructing GraphiteConfig{} literals.

Included is an example that makes paths for reported percentiles look more like
those used by twitter's ostrich library.